+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Change Log
 
-See [docs/CHANGELOG.rst](https://jax.readthedocs.io/en/latest/CHANGELOG.html).
+See [docs/CHANGELOG.md](https://jax.readthedocs.io/en/latest/CHANGELOG.html).